Refactor object cache to be separate entity and work without context
