+++ service/internal/cache/global.go
@@ -1,0 +1,4 @@
+package cache
+
+// Global cache that can be shared across the whole program.
+var Global = Scoped(DefaultExpiration)
