+++ service/internal/object/machinedeployment/accessor.go
@@ -12,20 +12,26 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/cluster-operator/service/controller/key"
+	"github.com/giantswarm/cluster-operator/service/internal/cache"
 	"github.com/giantswarm/cluster-operator/service/internal/object"
 )
 
 type accessor struct {
+	cache      cache.Interface
 	ctrlClient client.Client
 	logger     micrologger.Logger
 }
 
 type Config struct {
+	Cache      cache.Interface
 	CtrlClient client.Client
 	Logger     micrologger.Logger
 }
 
 func NewAccessor(config Config) (object.Accessor, error) {
+	if config.Cache == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Cache must not be empty", config)
+	}
 	if config.CtrlClient == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
 	}
@@ -34,6 +40,7 @@ func NewAccessor(config Config) (object.Accessor, error) {
 	}
 
 	a := &accessor{
+		cache:      config.Cache,
 		ctrlClient: config.CtrlClient,
 		logger:     config.Logger,
 	}
@@ -56,11 +63,9 @@ func (a *accessor) getAWSCluster(ctx context.Context, obj interface{}) (*infrast
 		return nil, microerror.Mask(err)
 	}
 
-	cache := object.CacheFromContext(ctx)
-
 	var awsCluster *infrastructurev1alpha2.AWSCluster
 	{
-		o, exists := cache.Get(clusterCacheKey(cr))
+		o, exists := a.cache.Get(clusterCacheKey(cr))
 		if exists {
 			awsCluster, err = toAWSCluster(o)
 			if err != nil {
@@ -77,7 +82,7 @@ func (a *accessor) getAWSCluster(ctx context.Context, obj interface{}) (*infrast
 				return nil, microerror.Mask(err)
 			}
 
-			cache.Put(clusterCacheKey(cr), awsCluster)
+			a.cache.Put(clusterCacheKey(cr), awsCluster)
 		}
 	}
